Repository misconfig fixed.

+++ main/src/main/java/com/ldg/main/Repository/AdCategoryRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ldg.main.Models.AdCategory;
 
-public interface AdCategoryRepository extends JpaRepository<AdCategory,Integer>{
+public interface AdCategoryRepository extends JpaRepository<AdCategory,Long>{
     
 }
