+++ main/src/main/java/com/ldg/main/Repository/AdRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ldg.main.Models.Ad;
 
-public interface AdRepository extends JpaRepository<Ad,Integer>{
+public interface AdRepository extends JpaRepository<Ad,Long>{
     
 }
